+++ packages/article/showcase-helper.js
@@ -16,7 +16,6 @@ import { sections } from "@times-components/storybook";
 import { scales, themeFactory } from "@times-components/ts-styleguide";
 import storybookReporter from "@times-components/tealium-utils";
 
-import { latestFromSection } from "./fixtures/full-article";
 import Article, { templates } from "./src/article";
 
 const preventDefaultedAction = decorateAction =>
@@ -226,7 +225,6 @@ const renderArticle = ({
   isTeaser,
   isMeteredExpired,
   additionalRelatedArticlesFlag,
-  latestFromSectionFlag,
   algoliaSearchKeys
 }) => (
   <ArticleProvider debounceTimeMs={0} id={id}>
@@ -288,10 +286,6 @@ const renderArticle = ({
             refetch={refetch}
             additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
             algoliaSearchKeys={algoliaSearchKeys}
-            latestFromSectionFlag={latestFromSectionFlag}
-            latestFromSection={latestFromSection.find(
-              ({ section: sectionName }) => sectionName === section
-            )}
             commentingConfig={commentingConfig}
           />
         </ContextProviderWithDefaults>
@@ -355,11 +349,7 @@ const renderArticleConfig = ({
     false,
     "User State"
   );
-  const latestFromSectionFlag = boolean(
-    "Latest from Section",
-    false,
-    "User State"
-  );
+
   return (
     <Fragment>
       {link}
@@ -392,7 +382,6 @@ const renderArticleConfig = ({
             section,
             template,
             additionalRelatedArticlesFlag,
-            latestFromSectionFlag,
             algoliaSearchKeys
           })}
         </ArticleConfigurator>