+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.8](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.8.7...@times-components/article-paragraph@1.8.8) (2022-09-06)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.8.7](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.8.6...@times-components/article-paragraph@1.8.7) (2022-09-06)
 
 **Note:** Version bump only for package @times-components/article-paragraph